requirements.txt
selenium
Appium-Python-Client
parameterized
requests
pymysql
pyexcel